Chat_with_MultiplePDF/requirements.txt
streamlit
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai